Added two frames for two players
Here we added two frames player1_frame and player2_frame.
The aim is to host images of a move corresponding to what
a user chooses. For instance, if player 1 chooses rock, we
show a rock in its corresponding frame and at the same time
if player2 chooses paper, we show an image of a paper at its
corresponding frame.